Moved from common back to root.
